admincrashboard/requirements.txt
python-pam==1.8.4
flask==2.0.3
flask-session==0.4.0
lxml==4.8.0
gunicorn
